app/requirements.txt
numpy
pandas
plotly-express
pyspark
streamlit